Remove redundant limit on number of search results
Accessing the value of `patientCount` using the not-null assertion
operator was resulting in a Null Pointer Exception if the count
operation had not yet completed. Limiting the number of search results
based on the count was not necessary anyway since the set of filters
used for the search was narrower than those used for the count which
meant that the count would always be greater than or equal to the number
of search results.