Introduce new `QuestionnaireAdapterItem` type
This is a sealed class that currently only has one subtype, wrapping
`QuestionnaireItemViewItem`.

More subtypes will be added in the future to support (for example)
repeated group headers, allowing us to break from the current 1:1
mapping of question:row in our adapters.

`QuestionnaireItemEditAdapter` and `QuestionnaireItemReviewAdapter` were
changed to use this new type.